# Use github.com/klauspost/compress/gzip, level=1