… LICENSES/vendor/github.com/klauspost/compress/LICENSE
… go.sum
… staging/src/k8s.io/apiserver/go.sum
… vendor/BUILD
… vendor/github.com/klauspost/compress/LICENSE
… vendor/github.com/klauspost/compress/flate/BUILD
… vendor/github.com/klauspost/compress/flate/deflate.go
… vendor/github.com/klauspost/compress/flate/dict_decoder.go
… vendor/github.com/klauspost/compress/flate/fast_encoder.go
… vendor/github.com/klauspost/compress/flate/huffman_bit_writer.go
… vendor/github.com/klauspost/compress/flate/huffman_code.go
… vendor/github.com/klauspost/compress/flate/huffman_sortByFreq.go
… vendor/github.com/klauspost/compress/flate/huffman_sortByLiteral.go
… vendor/github.com/klauspost/compress/flate/inflate.go
… vendor/github.com/klauspost/compress/flate/inflate_gen.go
… vendor/github.com/klauspost/compress/flate/level1.go
… vendor/github.com/klauspost/compress/flate/level2.go
… vendor/github.com/klauspost/compress/flate/level3.go
… vendor/github.com/klauspost/compress/flate/level4.go
… vendor/github.com/klauspost/compress/flate/level5.go
… vendor/github.com/klauspost/compress/flate/level6.go
… vendor/github.com/klauspost/compress/flate/regmask_amd64.go
… vendor/github.com/klauspost/compress/flate/regmask_other.go
… vendor/github.com/klauspost/compress/flate/stateless.go
… vendor/github.com/klauspost/compress/flate/token.go
… vendor/github.com/klauspost/compress/gzip/BUILD
… vendor/github.com/klauspost/compress/gzip/gunzip.go
… vendor/github.com/klauspost/compress/gzip/gzip.go
… vendor/modules.txt