#### What type of PR is this?

<!--
Add one of the following kinds:
/kind bug
/kind cleanup
/kind documentation
/kind design

Optionally add one or more of the following kinds if applicable:
/kind api-change
/kind deprecation
/kind failing-test
/kind flake
/kind regression
-->
/kind feature

#### What this PR does / why we need it:
This PR changes a gzip library and reduces level of compression to improve performance of gzip part in list calls. We see that gzipping is responsible for ~40% of time during large lists, when the network is fast.

I propose migrating to github.com/klauspost/compress/gzip library which has better performance. Comparison table:

The table below presents a time and output size for compressing list of 100k pods (Sizes before compression: PB: 250MiB, JSON: 392 MiB)

| Library                                         | PB output | PB time    | JSON output | JSON time |
|-------------------------------------------------|-----------|------------|-------------|-----------|
| compress/gzip, level=4 (baseline)               | 20.2MiB   | 1.14s      | 24.28MiB    | 1.79s     |
| compress/gzip, level=1                          | 25.3MiB   | 0.935s     | 37MiB       | 1.19s     |
| github.com/klauspost/compress/gzip, level=4     | 21MiB     | 0.63s      | 26MiB       | 1.02s     |
| **github.com/klauspost/compress/gzip, level=1** | **23MiB** | **0.404s** | **33MiB**   | **0.64s** |

While we see output size increase on level=1 from 24MiB -> 33MiB for JSON and 20.2MiB -> 23MiB for PB it is still very reasonable compression ratio (~10%, it was 6-8%) and the performance gain is significant (1.14s -> 0.404s for PB, 1.79s -> 0.64s for JSON).

Benchmark result:

```
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/endpoints/handlers/responsewriters
BenchmarkSerializeObject1000PodsPB
    writers_test.go:459: Payload size: 2629726, expected output size: 250981, ratio: 0.10
    writers_test.go:459: Payload size: 2629726, expected output size: 250508, ratio: 0.10
    writers_test.go:459: Payload size: 2629726, expected output size: 251379, ratio: 0.10
BenchmarkSerializeObject1000PodsPB-12        	     244	   5133290 ns/op	  635862 B/op	      43 allocs/op
BenchmarkSerializeObject10000PodsPB
    writers_test.go:459: Payload size: 26297133, expected output size: 2487867, ratio: 0.09
    writers_test.go:459: Payload size: 26297133, expected output size: 2487904, ratio: 0.09
BenchmarkSerializeObject10000PodsPB-12       	      24	  44078262 ns/op	 5037060 B/op	      48 allocs/op
BenchmarkSerializeObject100000PodsPB
    writers_test.go:459: Payload size: 262971335, expected output size: 24874023, ratio: 0.09
    writers_test.go:459: Payload size: 262971335, expected output size: 24866939, ratio: 0.09
    writers_test.go:459: Payload size: 262971335, expected output size: 24854519, ratio: 0.09
BenchmarkSerializeObject100000PodsPB-12      	       3	 404120836 ns/op	80080816 B/op	      59 allocs/op
BenchmarkSerializeObject1000PodsJSON
    writers_test.go:459: Payload size: 4118115, expected output size: 360565, ratio: 0.09
    writers_test.go:459: Payload size: 4118115, expected output size: 348033, ratio: 0.08
    writers_test.go:459: Payload size: 4118115, expected output size: 352384, ratio: 0.09
BenchmarkSerializeObject1000PodsJSON-12      	     139	   8611449 ns/op	 1262105 B/op	      45 allocs/op
BenchmarkSerializeObject10000PodsJSON
    writers_test.go:459: Payload size: 41180777, expected output size: 3546179, ratio: 0.09
    writers_test.go:459: Payload size: 41180777, expected output size: 3558888, ratio: 0.09
BenchmarkSerializeObject10000PodsJSON-12     	      13	  78464459 ns/op	10057800 B/op	      49 allocs/op
BenchmarkSerializeObject100000PodsJSON
    writers_test.go:459: Payload size: 411808348, expected output size: 35457492, ratio: 0.09
    writers_test.go:459: Payload size: 411808348, expected output size: 35453142, ratio: 0.09
BenchmarkSerializeObject100000PodsJSON-12    	       2	 645729630 ns/op	80271240 B/op	      63 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/endpoints/handlers/responsewriters	24.620s
```

#### Which issue(s) this PR fixes:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```

/assign @wojtek-t