@wojtek-t 2021-02-22
This looks like a reasonable direction to me, but I would like to hear from api-machinery folks too (in particular from outside Google).

/assign @smarterclayton @deads2k 

Clayton/David - Thoughts?