@dims 2021-02-22
@wojtek-t @mborsz - Existing clients should be able to decompress without issues right? ( they may end up consuming more memory? )