@wojtek-t 2021-02-22
> @wojtek-t @mborsz - Existing clients should be able to decompress without issues right? ( they may end up consuming more memory? )

That's my understanding, but yes - this needs a plan for validating it. But I wanted to first check check if this is something we will consider.