@smarterclayton 2021-02-24
This seems reasonable, I've seen a number of other ecosystem users running this library instead of the std.  I would like to see a rough description of the risk involved and the maturity and test level of the library, as well as a reasonable effort to ensure this is broadly interoperable, but other than that I'm +1 on direction.