@deads2k 2021-03-04
> I would like to see a rough description of the risk involved and the maturity and test level of the library, as well as a reasonable effort to ensure this is broadly interoperable, but other than that I'm +1 on direction.

Any way to be more concrete for the contributor.  I'm ok with the size/CPU tradeoff shown here, but it's not clear to me how to assess the risk and compatibility.