@BenTheElder 2021-06-04
This would need a rebase to drop the BUILD files (you can run hack/update-bazel.sh during interactive rebase to delete any build files automatically).

It also seems to still be pending more data / info.

@mborsz are you still interested in pursuing this?