@mborsz 2021-08-02
I still think that it make sense to continue this, but I don't have capacity to continue this PR. I opened an issue https://github.com/kubernetes/kubernetes/issues/104071 to continue this

/close